# internal/api/handlers/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LLM Pricing API",
    "version": "1.0.0",
    "description": "Reconciled, multi-source pricing data API for LLM models. Aggregates pricing from OpenRouter, LiteLLM, and provider docs, reconciles discrepancies, stores immutable price history, and serves it through a versioned REST API.",
    "contact": {
      "url": "https://llmpricing.dev"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Unkey API key. Include as: Authorization: Bearer <key>"
      }
    },
    "schemas": {
      "TrustMeta": {
        "type": "object",
        "description": "Trust metadata attached to every pricing response.",
        "properties": {
          "confirmed_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the price was most recently confirmed by reconciliation."
          },
          "source": {
            "type": "string",
            "description": "Human-readable name of the data source (e.g. 'OpenRouter', 'LiteLLM')."
          },
          "confidence": {
            "type": "string",
            "enum": ["high", "medium", "low"],
            "description": "Confidence level: high = 2+ sources agree, medium = single-source confirmed, low = unconfirmed."
          },
          "age_hours": {
            "type": "number",
            "format": "float",
            "description": "How many hours ago the price was last confirmed."
          },
          "change_velocity": {
            "type": "number",
            "format": "float",
            "description": "Number of price changes in the past 30 days."
          }
        },
        "required": ["confirmed_at", "source", "confidence", "age_hours", "change_velocity"]
      },
      "ModelResponse": {
        "type": "object",
        "description": "A single LLM model with its latest pricing.",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Internal model ID."
          },
          "provider": {
            "type": "string",
            "description": "Provider name (e.g. 'openai', 'anthropic')."
          },
          "name": {
            "type": "string",
            "description": "Human-readable model name."
          },
          "slug": {
            "type": "string",
            "description": "URL-safe model identifier."
          },
          "modality": {
            "type": "string",
            "description": "Model modality (e.g. 'text', 'multimodal')."
          },
          "context_window": {
            "type": ["integer", "null"],
            "description": "Context window size in tokens."
          },
          "price_input": {
            "type": "number",
            "format": "float",
            "description": "Input cost per token."
          },
          "price_output": {
            "type": "number",
            "format": "float",
            "description": "Output cost per token."
          }
        },
        "required": ["id", "provider", "name", "slug", "modality", "price_input", "price_output"]
      },
      "ModelEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ModelResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/TrustMeta"
          }
        },
        "required": ["data", "meta"]
      },
      "ModelListEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelResponse"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/TrustMeta"
          }
        },
        "required": ["data", "meta"]
      },
      "HistoryRow": {
        "type": "object",
        "description": "A single price history record.",
        "properties": {
          "input_cost_per_token": {
            "type": "number",
            "format": "float"
          },
          "output_cost_per_token": {
            "type": "number",
            "format": "float"
          },
          "source": {
            "type": "string"
          },
          "confirmed_at": {
            "type": "string",
            "format": "date-time"
          },
          "recorded_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["input_cost_per_token", "output_cost_per_token", "source", "confirmed_at", "recorded_at"]
      },
...
      "ProblemDetail": {
        "type": "object",
        "description": "RFC 7807 Problem Details for HTTP APIs.",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "description": "A URI reference that identifies the problem type."
          },
          "title": {
            "type": "string",
            "description": "A short, human-readable summary of the problem type."
          },
          "status": {
            "type": "integer",
            "description": "The HTTP status code."
          },
          "detail": {
            "type": "string",
            "description": "A human-readable explanation specific to this occurrence."
          },
          "instance": {
            "type": "string",
            "format": "uri",
            "description": "A URI reference that identifies the specific occurrence."
          }
        },
        "required": ["type", "title", "status"]
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid API key.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      },
      "Forbidden": {
        "description": "API key tier does not have access to this endpoint.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      },
      "NotFound": {
        "description": "Requested resource not found.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Rate limit exceeded.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/models": {
      "get": {
        "operationId": "listModels",
        "summary": "List models",
        "description": "Returns a paginated list of LLM models with current pricing. Available on the Free tier.",
        "tags": ["models"],
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "description": "Filter by provider name (case-insensitive).",
            "schema": { "type": "string" }
          },
          {
            "name": "modality",
            "in": "query",
            "description": "Filter by modality (e.g. 'text', 'multimodal').",
            "schema": { "type": "string" }
          },
          {
            "name": "min_context",
            "in": "query",
            "description": "Filter to models with context_window >= this value.",
            "schema": { "type": "integer" }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number (1-based).",
            "schema": { "type": "integer", "default": 1, "minimum": 1 }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Results per page (max 200).",
            "schema": { "type": "integer", "default": 50, "maximum": 200 }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated model list.",
            "headers": {
              "X-Total-Count": {
                "description": "Total number of models matching the filter.",
                "schema": { "type": "integer" }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListEnvelope"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/v1/models/{id}": {
      "get": {
        "operationId": "getModel",
        "summary": "Get model by ID",
        "description": "Returns a single LLM model with its current pricing and trust metadata. Available on the Free tier.",
        "tags": ["models"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Model integer ID.",
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Model detail.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelEnvelope"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/v1/models/{id}/history": {
      "get": {
        "operationId": "getModelHistory",
        "summary": "Get price history for a model",
        "description": "Returns the full price history for a model. Requires Developer tier or above.",
        "tags": ["models"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Model integer ID.",
            "schema": { "type": "integer" }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Start of time range (ISO 8601).",
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "to",
            "in": "query",
            "description": "End of time range (ISO 8601).",
            "schema": { "type": "string", "format": "date-time" }
          }
        ],
        "responses": {
          "200": {
            "description": "Price history records.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/HistoryRow" }
                    },
                    "meta": { "$ref": "#/components/schemas/TrustMeta" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/v1/compare": {
      "get": {
        "operationId": "compareModels",
        "summary": "Compare models",
        "description": "Returns pricing and trust metadata for up to 5 models side-by-side. Available on the Free tier.",
        "tags": ["models"],
        "parameters": [
          {
            "name": "models",
            "in": "query",
            "required": true,
            "description": "Comma-separated list of model IDs (max 5).",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Comparison results.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (e.g. too many models, invalid IDs).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetail" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/v1/recommend": {
      "get": {
        "operationId": "recommendModels",
        "summary": "Recommend models",
        "description": "Returns models ranked by price for a given task and context size. Requires Developer tier or above.",
        "tags": ["models"],
        "parameters": [
          {
            "name": "task",
            "in": "query",
            "description": "Task type (e.g. 'summarization', 'code-generation').",
            "schema": { "type": "string" }
          },
          {
            "name": "context_size",
            "in": "query",
            "description": "Minimum context window required (tokens).",
            "schema": { "type": "integer" }
          },
          {
            "name": "max_price_input",
            "in": "query",
            "description": "Maximum input cost per token.",
            "schema": { "type": "number", "format": "float" }
          }
        ],
        "responses": {
          "200": {
            "description": "Ranked model recommendations.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListEnvelope"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
...
}